Commented the folder creation at project root level
The files are project root are not shown (i.e. README files). Therefore, if you make a folder, it doesn't appear in the UI. The functionality is tested and works, but the scanning of root-level files should be implemented first (more importantly, a descision should be taken which files to show).